test: some tests
